+++ plugin-dev/Source/Sentry/Private/SentryScope.cpp
@@ -43,20 +43,28 @@ void USentryScope::ClearAttachments()
 	NativeImpl->ClearAttachments();
 }
 
-void USentryScope::SetTagValue(const FString& Key, const FString& Value)
+void USentryScope::SetTag(const FString& Key, const FString& Value)
 {
 	if (!NativeImpl)
 		return;
 
-	NativeImpl->SetTagValue(Key, Value);
+	NativeImpl->SetTag(Key, Value);
 }
 
-FString USentryScope::GetTagValue(const FString& Key) const
+FString USentryScope::GetTag(const FString& Key) const
 {
 	if (!NativeImpl)
 		return FString();
 
-	return NativeImpl->GetTagValue(Key);
+	return NativeImpl->GetTag(Key);
+}
+
+bool USentryScope::TryGetTag(const FString& Key, FString& Value) const
+{
+	if (!NativeImpl)
+		return false;
+
+	return NativeImpl->TryGetTag(Key, Value);
 }
 
 void USentryScope::RemoveTag(const FString& Key)
@@ -131,20 +139,28 @@ void USentryScope::RemoveContext(const FString& Key)
 	NativeImpl->RemoveContext(Key);
 }
 
-void USentryScope::SetExtraValue(const FString& Key, const FString& Value)
+void USentryScope::SetExtra(const FString& Key, const FString& Value)
 {
 	if (!NativeImpl)
 		return;
 
-	NativeImpl->SetExtraValue(Key, Value);
+	NativeImpl->SetExtra(Key, Value);
 }
 
-FString USentryScope::GetExtraValue(const FString& Key) const
+FString USentryScope::GetExtra(const FString& Key) const
 {
 	if (!NativeImpl)
 		return FString();
 
-	return NativeImpl->GetExtraValue(Key);
+	return NativeImpl->GetExtra(Key);
+}
+
+bool USentryScope::TryGetExtra(const FString& Key, FSentryVariant& Value) const
+{
+	if (!NativeImpl)
+		return false;
+
+	return NativeImpl->TryGetExtra(Key, Value);
 }
 
 void USentryScope::RemoveExtra(const FString& Key)